Item Scaling Enhancements
Changes the rendering source for items from "ground" to "head."
This allows for mod devs/modpack authors to set a custom item scaling and position for use in Backslot within the "head" section of an item's .json.
Instead of using "ground," now you can make changes without affecting how it looks when dropped, and 99.9% of items that would go in a backslot or beltslot won't be able to be put into the head armor slot, so it's a really good place to do so without writing a bunch of new code.
Also adds a chestplate check and moves items to be flush with the player instead of floating when you aren't wearing a chestplate.
The minecraft assests, handheld and crossbow.json are also required to fix vanilla items due to the rendering changes.
The only addition to these files from vanilla is the usage of the "head" category, that is not in use by default for these files.